Refactor more measurements to decimal. Update tests.

+++ src/ShapeCrawler/ShapeCollection/AutoShape.cs
@@ -92,7 +92,7 @@ internal sealed class AutoShape : CopyableShape
         {
             Color = skColorOutline,
             IsAntialias = true,
-            StrokeWidth = UnitConverter.PointToPixel(this.Outline.Weight),
+            StrokeWidth = (float)UnitConverter.PointToPixel(this.Outline.Weight),
             Style = SKPaintStyle.Stroke
         };
